feat: implement token validation in reset password flow and clean up ResetPassword component
