+++ src/components/modules/auth/ui/pages/Login.tsx
@@ -4,7 +4,7 @@ import { LoginForm } from "../forms/LoginForm";
 
 export const Login = () => {
   return (
-    <div className="flex min-h-svh flex-col items-center justify-center bg-muted p-6 md:p-10">
+    <div className="flex min-h-svh flex-col items-center justify-center p-6 md:p-10">
       <div className="w-full max-w-sm md:max-w-xl">
         <LoginForm />
       </div>
